fix ci unit test isolation
