+++ packages/database/src/flashcard-data.ts
@@ -30,16 +30,15 @@ import {
   flashcardSetEnrollment,
   workspace,
 } from "./schema";
-import { canonicalizeLearningTaxonomy } from "./learning-taxonomy";
+import {
+  normalizeFlashcardTaxonomy,
+  type FlashcardTaxonomy,
+} from "./learning-taxonomy";
 
 export type FlashcardSourceType = "manual" | "ai-generated";
 export type FlashcardCardKind = "flashcard" | "multiple_choice_quiz";
 export type FlashcardEnrollmentStatus = "active" | "paused";
-export interface FlashcardTaxonomy {
-  concept: string;
-  subject: string;
-  topic: string;
-}
+export type { FlashcardTaxonomy };
 export type FlashcardDisplayState =
   | "new"
   | "learning"
@@ -219,45 +218,7 @@ function sanitizeTags(value: string[] | undefined | null) {
     .slice(0, 12);
 }
 
-function sanitizeTaxonomyField(
-  value: unknown,
-  fieldName: keyof FlashcardTaxonomy
-) {
-  if (typeof value !== "string") {
-    return null;
-  }
-
-  const trimmed = value.trim();
-  if (!trimmed) {
-    return null;
-  }
-
-  return trimmed.slice(0, fieldName === "concept" ? 180 : 120);
-}
-
-export function normalizeFlashcardTaxonomy(
-  value: unknown
-): FlashcardTaxonomy | null {
-  if (!(value && typeof value === "object" && !Array.isArray(value))) {
-    return null;
-  }
-
-  const record = value as Record<string, unknown>;
-  const subject = sanitizeTaxonomyField(record.subject, "subject");
-  const topic = sanitizeTaxonomyField(record.topic, "topic");
-  const concept = sanitizeTaxonomyField(record.concept, "concept");
-
-  if (!(subject && topic && concept)) {
-    return null;
-  }
-
-  return canonicalizeLearningTaxonomy({
-    concept,
-    subject,
-    text: [subject, topic, concept].join(" "),
-    topic,
-  });
-}
+export { normalizeFlashcardTaxonomy };
 
 export function assertFlashcardTaxonomy(
   value: unknown,
