+++ packages/database/src/learning-taxonomy.ts
@@ -4,6 +4,8 @@ export interface LearningTaxonomy {
   topic: string;
 }
 
+export type FlashcardTaxonomy = LearningTaxonomy;
+
 function normalizeWhitespace(value: string) {
   return value.replace(/[_-]+/g, " ").replace(/\s+/g, " ").trim();
 }
@@ -61,4 +63,44 @@ export function canonicalizeLearningTaxonomy(input: {
     subject,
     topic,
   };
+}
+
+function sanitizeTaxonomyField(
+  value: unknown,
+  fieldName: keyof FlashcardTaxonomy
+) {
+  if (typeof value !== "string") {
+    return null;
+  }
+
+  const trimmed = value.trim();
+  if (!trimmed) {
+    return null;
+  }
+
+  return trimmed.slice(0, fieldName === "concept" ? 180 : 120);
+}
+
+export function normalizeFlashcardTaxonomy(
+  value: unknown
+): FlashcardTaxonomy | null {
+  if (!(value && typeof value === "object" && !Array.isArray(value))) {
+    return null;
+  }
+
+  const record = value as Record<string, unknown>;
+  const subject = sanitizeTaxonomyField(record.subject, "subject");
+  const topic = sanitizeTaxonomyField(record.topic, "topic");
+  const concept = sanitizeTaxonomyField(record.concept, "concept");
+
+  if (!(subject && topic && concept)) {
+    return null;
+  }
+
+  return canonicalizeLearningTaxonomy({
+    concept,
+    subject,
+    text: [subject, topic, concept].join(" "),
+    topic,
+  });
 }
